tracker: Dump entire collection by default
This is kinda slow, which we should fix by paging the queries a bit so
that it starts printing stuff to stdout after querying the first couple
of hundred rows.

Also changes to how results are printed. The idea of 'result tables' is
codified better now for things that can't be represented as collections
(i.e. the 'number of songs per artist' table).