sync: Improve track renumbering
This now works for playlists as well as collections with nested albums.

It also avoids situations like this:

    001 - Track 1.mp3
    003 - Track 3.mp3
    02 - Track 2.mp3

Previously if a file had a number that was numerically equal to the
string that calliope-sync was going to prepend, it would leave it as
is. But the asciibetical sorting done by most MP3 players means that we
need to be exact in the number of zeros we prepend!