tracker: Add support for app domains, split up commands
There are some things that are not possible using the session-wide
Tracker instance, such as controlling which locations are scanned
using the commandline. This commit adds support for a feature I hacked
together called 'app domains' which allows using Tracker as if it was
a library, with data stored separately to the user's session-wide
instance.

I split up the various commands to make things more logical.