subprojects: Update tracker-app-domain to fix hang
We can now use it multiple times in the same process; which means we
could at some point switch back to running tests using
click.testing.CliRunner. However there are other issues with that such
as lack of separation of stdout vs stderr so I will stick with the
current approach for now.