tracker: Replace `show` with `local-{artists,tracks,albums}`
The `show` command would always return a structured list grouped by
artist and then album. This can be useful but sometimes we just want a
flat list of tracks or a flat list of artists.

This is now possible, the `show` command has become `local-albums` and
the new `local-tracks` and `local-albums` command can produce a flat list
of tracks or artists, respectively.