Remove JsonCache implementation
The GdbmCache is more efficient overall, and can support concurrent
readers and writers.

Note that in order to support concurrent readers and writers, we reload
the GDBM database each time we want to look up something in the cache.
This is inefficient compared to opening the database once; however the
code is much simpler this way and the performance penalty is not enough
to cause problems right now.

Performance comparison:

  - JsonCache: mean read speed 6.4748 μs , mean write speed 6,000μs
  - GdbmCache: mean read speed 133μs, mean write speed 118μs