lastfm: Handle multiple scrobbles with same datetime
This shouldn't happen, but it does. We must treat these entries as
individual scrobbles, because we decide whether we are fully synced
by comparing how many stored scrobbles we have to the total number
the server sent. If we don't store everything that the server sends
us then we'll always try to fetch more.