Separate all CLI related code into a single 'cli.py' file
This is a reorganization of the code which aims to make it possible to
build the CLI documentation with a minimal set of dependencies.

It also paves the way for implementing a real Python API to complement
the current CLI, although further work will be needed for that.

The main goal of this reorganization is to enable auto-generation of
reference documentation on readthedocs.org. The sphinx-click plugin
needs to be able to do `import calliope.cli`. It's now possible to
do this without needing all of Calliope's dependencies to be installed
on the system, which makes it practical to do this on a readthedocs.org
runner.