Replace 'dateparser' with 'parsedatetime'
The latter appears to have much poorer internationalization, so we might
switch back in future. However we are a long way off from having the
tools accept their commandline arguments in languages other than English
at the moment. Meanwhile, 'dateparser' has some requirements that are
difficult to satisfy on container OSes: the 'regex' module isn't
packaged in Alpine Linux and has C code bundled with it, so `pip
install` needs a full C toolchain available to be able to install this
module. Right now it's not worth that pain.